Move power event watcher to its own package
